test(transform): split aggregate escape regressions
